requirements.txt
asgiref==3.4.0
boto3==1.17.102
botocore==1.20.102
Django==3.2.4
django-countries==7.2.1
django-phonenumber-field==5.2.0
django-storages==1.11.1
djangorestframework==3.12.4
djongo==1.3.6
dnspython==1.16.0
gunicorn==20.1.0
jmespath==0.10.0
phonenumbers==8.12.26
Pillow==8.2.0
pymongo==3.11.4
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.6
uuid==1.30
whitenoise==5.2.0
